Move NextBytesMatchAsync to StreamExtensions class
