+++ src/NuGetGallery.Services/Helpers/StreamHelper.cs
@@ -101,45 +101,6 @@ namespace NuGetGallery.Services.Helpers
             }
         }
 
-        /// <summary>
-        /// Checks whether stream's next bytes are the same as <paramref name="expectedBytes"/>.
-        /// </summary>
-        /// <param name="stream">Stream to read from.</param>
-        /// <param name="expectedBytes">Expected bytes.</param>
-        /// <returns>True if it manages to read the same amount of bytes as in <paramref name="expectedBytes"/> and 
-        /// the bytes read are the same as in <paramref name="expectedBytes"/> array, false otherwise.</returns>
-        /// <remarks>
-        /// The method assumes that stream is properly positioned before calling by the caller. It will not seek anywhere.
-        /// If the length of <paramref name="expectedBytes"/> is 0, method returns true.
-        /// </remarks>
-        public static async Task<bool> NextBytesMatchAsync(this Stream stream, byte[] expectedBytes)
-        {
-            if (stream == null)
-            {
-                throw new ArgumentNullException(nameof(stream));
-            }
-
-            if (expectedBytes == null)
-            {
-                throw new ArgumentNullException(nameof(expectedBytes));
-            }
-
-            if (expectedBytes.Length == 0)
-            {
-                return true;
-            }
-
-            var actualBytes = new byte[expectedBytes.Length];
-            var bytesRead = await stream.ReadAsync(actualBytes, 0, actualBytes.Length);
-
-            if (bytesRead != expectedBytes.Length)
-            {
-                return false;
-            }
-
-            return expectedBytes.SequenceEqual(actualBytes);
-        }
-
         private static int GetNeededBytesToRead(int totalBytesRead, int maxSize)
         {
             var neededReadSize = maxSize - totalBytesRead + 1;
